refactor: Dto 클래스 수정
1. 의존성 문제로 인해 order-api 모듈 작성에 어려움을 느껴 SellerDto.java, UserDto.java를 Game-domain으로 옮김
