+++ Game-domain/src/main/java/com/user/gamedomain/domain/Dto/SellerDto.java
@@ -1,0 +1,16 @@
+package com.user.gamedomain.domain.Dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+@AllArgsConstructor
+public class SellerDto {
+
+  private String company;
+  private String bank;
+  private String number;
+
+}
