+++ Game-domain/src/main/java/com/user/gamedomain/domain/Dto/UserDto.java
@@ -1,0 +1,15 @@
+package com.user.gamedomain.domain.Dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+@AllArgsConstructor
+public class UserDto {
+
+  private Long id;
+  private String email;
+
+}
